Don't write to streaming clients when we are paused.
This was causing clients that connected to the stream while the player was
paused to get sent the buffer of the last played audio. Quite uselessly
because there is no reason to full up their cache if they are just going to go
ahead and empty it again.

Maybe the cache of recent audio should be flushed on seek too?